Refactor logic
pr:refactor
