+++ src/commands/mod.rs
@@ -1,0 +1,13 @@
+pub mod fix_chain;
+pub mod list;
+pub mod merge;
+pub mod prep;
+pub mod restack;
+pub mod update;
+
+pub use fix_chain::fix_chain;
+pub use list::list_prs_display;
+pub use merge::merge_prs_until;
+pub use prep::prep_squash;
+pub use restack::restack_existing;
+pub use update::build_from_tags;
